pyproject.toml
[tool.poetry]
name = "fedcloud-monitoring-tools"
version = "0.2.1"
description = "Monitoring fedcloud VMs and sites"
authors = ["Giuseppe La Rocca <giuseppe.larocca@egi.eu>",
	   "Sebastián Luna-Valero <sebastian.luna.valero@egi.eu>",
           "Enol Fernández <enol.fernandez@egi.eu>"]
readme = "README.md"
include = [
    { path = "fedcloud_monitoring_tools/data", format = ["sdist", "wheel"] }
]

[tool.poetry.scripts]
fedcloud-vo-monitor = "fedcloud_monitoring_tools.vm_monitor_cli:main"
fedcloud-sla-monitor = "fedcloud_monitoring_tools.sla_monitor_cli:main"
fedcloud-vo-testing = "fedcloud_monitoring_tools.vo_mon_cli:main"

[tool.poetry.dependencies]
python = "^3.12"
fedcloudclient = "^1.4.4"
ldap3 = "^2.9.1"
python-dateutil = "^2.9.0.post0"
paramiko = "^3.4.0"
IM-client = "^1.8.0"
httpx = "^0.27.2"
xmltodict = "^0.14.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
